requirements.txt
absl-py==2.1.0
aiohttp==3.9.3
aiosignal==1.3.1
asttokens==2.4.1
astunparse==1.6.3
attrs==23.2.0
beautifulsoup4==4.12.3
bleach==6.1.0
blinker==1.7.0
cachetools==5.3.2
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
comm==0.2.1
contourpy==1.2.0
cssutils==2.9.0
cycler==0.12.1
dataframe-image==0.2.3
debugpy==1.8.1
decorator==5.1.1
defusedxml==0.7.1
executing==2.0.1
fastjsonschema==2.19.1
Flask==3.0.2
flatbuffers==23.5.26
fonttools==4.48.1
frozenlist==1.4.1
gast==0.5.4
google-auth==2.28.0
google-auth-oauthlib==1.2.0
google-pasta==0.2.0
grpcio==1.60.1
h5py==3.10.0
html2image==2.0.4.3
idna==3.6
ipykernel==6.29.2
ipython==8.21.0
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.3
joblib==1.3.2
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter_client==8.6.0
jupyter_core==5.7.1
jupyterlab_pygments==0.3.0
keras==2.15.0
kiwisolver==1.4.5
lazypredict==0.2.12
libclang==16.0.6
lightgbm==4.3.0
lxml==5.1.0
Markdown==3.5.2
MarkupSafe==2.1.5
matplotlib==3.8.2
matplotlib-inline==0.1.6
mistune==3.0.2
ml-dtypes==0.2.0
multidict==6.0.5
nbclient==0.9.0
nbconvert==7.16.0
nbformat==5.9.2
nest-asyncio==1.6.0
numpy==1.26.4
oauthlib==3.2.2
opencv-python==4.9.0.80
opt-einsum==3.3.0
packaging==23.2
pandas==2.2.0
pandocfilters==1.5.1
parso==0.8.3
pillow==10.2.0
platformdirs==4.2.0
prompt-toolkit==3.0.43
protobuf==4.25.3
psutil==5.9.8
pure-eval==0.2.2
pyasn1==0.5.1
pyasn1-modules==0.3.0
Pygments==2.17.2
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2024.1
pywin32==306
pyzmq==25.1.2
referencing==0.33.0
requests==2.31.0
requests-oauthlib==1.3.1
rpds-py==0.18.0
rsa==4.9
scikit-learn==1.4.0
scipy==1.12.0
seaborn==0.13.2
six==1.16.0
soupsieve==2.5
stack-data==0.6.3
tensorboard==2.15.2
tensorboard-data-server==0.7.2
tensorflow==2.15.0
tensorflow-estimator==2.15.0
tensorflow-intel==2.15.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.4.0
threadpoolctl==3.2.0
tinycss2==1.2.1
tornado==6.4
tqdm==4.66.2
traitlets==5.14.1
typing_extensions==4.9.0
tzdata==2024.1
urllib3==2.2.0
wcwidth==0.2.13
webencodings==0.5.1
websocket-client==1.7.0
Werkzeug==3.0.1
wrapt==1.14.1
xgboost==2.0.3
yarl==1.9.4
